backend/requirements.txt
flask
gunicorn
google-cloud-storage
google-cloud-speech
google-cloud-bigquery
librosa
noisereduce
soundfile
webrtcvad-wheels
numpy